Split base and provided containers.
